+++ android/src/main/java/ai/nexconn/ai_nexconn_chatui_plugin/VersionUpload.java
@@ -2,6 +2,6 @@ package ai.nexconn.ai_nexconn_chatui_plugin;
 
 public class VersionUpload {
     public static String getVersion() {
-        return "26.2.7";
+        return "26.2.8";
     }
 }